+++ lib/cloud-metadata/callback-coordination.js
