+++ lib/cloud-metadata/index.js
@@ -2,7 +2,7 @@
 const { getMetadataAwsV1, getMetadataAwsV2 } = require('./aws')
 const { getMetadataAzure } = require('./azure')
 const { getMetadataGcp } = require('./gcp')
-const { CallbackCoordination } = require('./fetch-coordination')
+const { CallbackCoordination } = require('./callback-coordination')
 const DEFAULT_PROVIDER_CONFIG = {
   aws: {
     host: '169.254.169.254',
